fastapi_app/requirements.txt
black==22.3.0
fastapi>=0.68.0,<0.69.0
isort==5.10.1
mypy==0.950
opentelemetry-api==1.11.1
opentelemetry-exporter-otlp==1.11.1
opentelemetry-instrumentation-fastapi==0.30b1
opentelemetry-sdk==1.11.1
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
